range 값 적용, 디자인 수정
